[build] use RegisterTaskObject in <GitCommitInfo/> (#2811)
When building `Xamarin.Android.sln` (even with no changes),
`<GitCommitInfo/>` runs 15 times, which takes about 6 seconds on my
machine.

Reviewing the input values to `<GitCommitInfo/>`, 2/3 of the calls
seem to be duplicate. This is a classic example where
`RegisterTaskObject()` can help.

I cached the `CommitInfo` output property using the important inputs
as a key to `RegisterTaskObject()`:

  * `GitRemoteName`
  * `SubmoduleName`
  * `WorkingDirectory`
  * `XASourceDirectory`

I used the simple syntax for creating a C# tuple as the lookup key.
I had to update the TFV of two projects to v4.7 to use the simplified
tuple syntax.  I prefer this over
`Tuple<string,string,string,string>`.

This saves ~4 seconds on incremental builds of `Xamarin.Android.sln`.