[build] improve XA's build for proguard & r8 (#2809)
If you build `Xamarin.Android.sln`, a lot of stuff happens even if
there are no changes...

Our internal build targets don't always follow our own
"MSBuild Best Practices", and thus we have a few MSBuild targets that
always run.

~~ `r8.csproj` ~~

The main problem here was that we were invoking `gradlew` every time,
and it takes quite a while to startup when using the `--no-daemon`
switch.

I added proper `Inputs` and `Outputs` for the `_BuildR8` MSBuild
target, as well as a `<Touch/>` call.

These changes saved 6 seconds(!!!) on incremental builds of
`Xamarin.Android.sln`.


~~ `proguard.csproj` ~~

This project was invoking `<Ant/>` and copying files every time.

I cleaned this project up a bit by creating an `@(_Outputs)` item
group to be used throughout.  Adding `Inputs` and `Outputs` to the
`_BuildProGuard` target and a `<Touch/>` fixed the incremental build.

The `Clean` target was actually broken.  Since there is a chain of
`@(ProjectReference)` here, the
`Xamarin.Android.Tools.BootstrapTasks.dll` assembly which contains
the `<Ant/>` MSBuild task gets removed!  Then the actual
call to `<Ant/>` fails due to the assembly being missing.  The fix
here is to just completely replace the `Clean` target with our own,
so `@(ProjectReference)` will not be cleaned.  This is the same
behavior `r8.csproj` already had.

There was also a general problem with the `CoreCompile` target.  I
don't think it is appropriate to run `<Ant/>` during a design-time
build in Visual Studio.  I just made this target empty--I think it is
here to appease OmniSharp.

These changes saved ~2 seconds on incremental builds of
`Xamarin.Android.sln`.

This brings us to a grand total of ~8 seconds saved.