[Xamarin.Android.Build.Tasks] Add custom meta data to @(ReferencePath)
As part of ongoing work with the support libraries
we will want to start adding custom metadata to
the assembly references.

This data will inform the build process that it
can skip processing of certain items. Or remove
that assembly completely from the final package.

We did some initial work on this in commit 02c07ed6.
However that was strickly limited to the
`ConvertResourcesCases` task. We now need a more
generic solution.

In this case we have a new Task

	`AppendCustomMetadataToItemGroup`

This task takes two item groups and will copy over
the metadata between the two. This is based on
a filename comparision. The output will be
the `Inputs` list but with any additional metadata
added to the appropriate items. We can then use
this list in the build process and qurey the
metadata as needed.

The `MetaDataItems` list can contain a number of
different ItemGroup with different metadata. If
this is the case the result will be a union of the
two.

```
<ItemGroup>
  <I Include="Foo" />
  <A Include="Foo">
    <Meta1>True</Meta1>
  </A>
  <B Include="Foo">
    <Meta2>False</Meta2>
  </B>
</ItemGroup
```

Given the above the result will be as follows
when `@(A);@(B)` are passed as the `MetaDataItems`

```
<ItemGroup>
  <I Include="Foo>
    <Meta1>True</Meta1>
    <Meta2>False</Meta2>
  </I>
</ItemGroup>
```