[tools] Download aapt2 and bundletool from google. (#2763)
In order to support things like `AppBundle` we cannot use the version
of `aapt2` that is shipped with the Android SDK.  Google recommends
[downloading it from maven][1].

If we are shipping our own version of `aapt2` then we no longer need
to support using the Android SDK version.  If a user really wants to
use the Android SDK version they can set the `$(Aapt2ToolPath)`
MSBuild property manually.

We should also download `bundletool` and ship that too.  This will
mean we can rely on both of these being available when we start
looking at `AppBundle` support.

[1]: https://developer.android.com/studio/build/building-cmdline#download_aapt