[Xamarin.Android.Build.Tasks] prepend to $(_ResolveMonoAndroidSdksDependsOn)
Downstream in monodroid, we are having an evaluation ordering problem
with our MSBuild targets. We were intending to append a target to
`$(_ResolveMonoAndroidSdksDependsOn)`, but in fact the opposite is
happening:

    Property reassignment: $(_ResolveMonoAndroidSdksDependsOn)="
            GetReferenceAssemblyPaths;
        " (previous value: "

            ;_SetupInstantRun
        ") at Xamarin.Android.Common.targets (868,2)

xamarin-android is overwriting what monodroid has for the value!

The behavior I'm seeing in MSBuild:

* `Xamarin.Android.Common.targets` is loaded
* All `<Import/>` in `Xamarin.Android.Common.targets` are loaded *and*
  `$(_ResolveMonoAndroidSdksDependsOn)` is evaluated
* `$(_ResolveMonoAndroidSdksDependsOn)` in
  `Xamarin.Android.Common.targets` is evaluated

I am not sure why MSBuild evaluates files in this way, but here we are.