[Xamarin.Android.Build.Tools] Bump android Tools versions.
Fixes #2433

We should be using the latest versions of the android tooling
if possible. If we can we should try to match the major version
with the max API level we support. In this case we are on Android
P, API Level 28. So we should be using the 28.x.x versions of
the tooling.

Since not all of the tools have a 28.x.x version, in those cases
we should just use the very latest.