spec & prototype of Android App Bundles
Context: https://github.com/xamarin/xamarin-android/issues/2727

This is a prototype that gets us this far:

* We generate a `.aab` file
* We can generate a `.apks` file specific for an attached device
* We can *install* the `.apks` file
* The app starts successfully!

This workflow is achieved by:

* The `<Aapt2Link/>` MSBuild task needs to pass `--proto-format`.
* The `<BuildBaseAppBundle/>` and `<BuildAppBundle/>` MSBuild tasks
  run instead of `<BuildApk/>`.
* The `<BuildApkSet/>` and `<InstallApkSet/>` tasks run instead of
  `adb install`. These are somewhat odd, but they use the attached
  device to decide which format APK set is needed. Otherwise the APK
  set was 200MB!

Some notes about Android App Bundles:

* App bundles use `android:extractNativeLibs="false"`, unless the
  target device's API level is too low.
* `$(AndroidUseAapt2)` is required.
* `$(EmbedAssembliesIntoApk)` is required.
* `$(_EmbeddedDSOsEnabled)` is required, regardless of
  `android:extractNativeLibs` value in `AndroidManifest.xml`.
* `$(AndroidUseApkSigner)` is turned off.
* `$(AndroidUseSharedRuntime)` is turned off.

~~ MonoRuntimeProvider ~~

Some changes were needed in `MonoRuntimeProvider.Bundled.java` in
java.interop to move some Java source files to this repo:

https://github.com/xamarin/java.interop/pull/422

We need the "split apks" to be in the list of APKs by calling
`ApplicationInfo.splitPublicSourceDirs`:

https://developer.android.com/reference/android/content/pm/ApplicationInfo.html#splitPublicSourceDirs

These are supported on API 21 and higher, so I had to create 4 `.java`
files:

* `MonoRuntimeProvider.Bundled.java`
* `MonoRuntimeProvider.Bundled.20.java`
* `MonoRuntimeProvider.Shared.java`
* `MonoRuntimeProvider.Shared.20.java`

I also extracted `MonoRuntimeProvider.Shared*.java` in
`_AddStaticResources`. We should remove the equivalent in monodroid,
to keep the code that *extracts* this file in the same repo.

I also removed `Xamarin.Android.Platform`, which is a dead package we
don't need code for anymore.