[build] support for building with VS 2019 (#2666)
Fixes: https://github.com/xamarin/xamarin-android/issues/2627

There are several things broken when building with MSBuild 16.0
and/or using [Visual Studio 2019][0].


~~ TargetFrameworkVersion ~~

Anything referencing MSBuild assemblies must now use
`$(TargetFrameworkVersion)`=v4.7.2, since the MSBuild assemblies are
now targeting v4.7.2.

We should take this as an opportunity to consolidate all
`$(TargetFrameworkVersion)` values, so I moved this property into
`Configuration.props`.  I was able to remove the property in many
projects, and the IDE still seemed to work fine.

The only hang up is that some versions of Mono don't have v4.7.2 yet.
I had to use the following trick, which use
[`ToolLocationHelper.GetPathToStandardLibraries()`][1]:

	<_StandardLibraryPath>$([Microsoft.Build.Utilities.ToolLocationHelper]::GetPathToStandardLibraries('.NETFramework', 'v4.7.2', ''))</_StandardLibraryPath>
	<TargetFrameworkVersion Condition=" '$(TargetFrameworkVersion)' == '' And '$(_StandardLibraryPath)' != '' ">v4.7.2</TargetFrameworkVersion>
	<TargetFrameworkVersion Condition=" '$(TargetFrameworkVersion)' == '' ">v4.7.1</TargetFrameworkVersion>

This allows us to fallback to v4.7.1 if we have to.


~~ Prepend<T> ~~

With `$(TargetFrameworkVersion)` changing, this method:

	static IEnumerable<T> Prepend<T> (this IEnumerable<T> l, T another)

Now exists in the BCL, so we can remove it.


~~ MSBuild.exe location ~~

In VS 2017, MSBuild is located at:

	%VsInstallDir%\MSBuild\15.0\Bin\MSBuild.exe

In VS 2019, it is now located in:

	%VsInstallDir%\MSBuild\Current\Bin\MSBuild.exe

Right now we have a bit of code that "finds" Visual Studio, MSBuild,
etc. so we should take this opportunity to improve it.

We originally thought about using MSBuildLocator:

	https://www.nuget.org/packages/Microsoft.Build.Locator/
	https://github.com/Microsoft/MSBuildLocator

But the licensing of the library was concerning... We were uncertain
if we could redistribute MSBuildLocator as part of an OSS product.

For now we can just execute `vswhere.exe`:

	https://github.com/Microsoft/vswhere/wiki/Find-MSBuild

I made a simple `vswhere.csproj` we can reference where this
functionality is needed.  Currently `xabuild` and
`Xamarin.ProjectTools` need to locate the Visual Studio directory.


~~ Other breakage in xabuild.exe ~~

In VS 2019, the path to Roslyn is a bit odd:

	%VsInstallDir%\MSBuild\15.0\Bin\Roslyn

I had to rework things to still work when combined with a different
MSBuild location:

	%VsInstallDir%\MSBuild\Current\Bin\MSBuild.exe

Additionally, our binding redirects weren't working in this project at
all.  `App.config` was *not* in the csproj!


~~ NUnit ~~

There does not appear to be an NUnit extension available for VS 2019
yet.  I frequently use the `Test Explorer` in VS to individually run
tests.

However, it looks like they are switching to a different model for
test frameworks.  Each test framework ships its own "adapter" that
enables the testing UIs in Visual Studio.  These are shipped on
NuGet, and so they can work without installing any extra extensions.

So we just need this NuGet package:

	https://www.nuget.org/packages/NUnit3TestAdapter/

This package should not affect anything on non-Windows platforms.


~~ ILRepack ~~

With the usage of `$(TargetFrameworkVersion)`=v4.7.2 we were hitting:

	"src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj" (default target) (1) ->
	    (ILRepacker target) ->
	    src/Xamarin.Android.Build.Tasks/ILRepack.targets(24,3): error : Failed to resolve assembly: 'netstandard, Version=2.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51' [src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj]

We need to specify `$(TargetPlatformDirectory)`; path logic in
`<ILRepack/>` appears to be failing, see:

	https://github.com/Alexx999/ILRepack.Lib.MSBuild.Task/blob/72c4c3bd051a166f5e3316b55457b24034c69a1d/ILRepack.Lib.MSBuild/ILRepack.cs#L148-L151
	https://github.com/Alexx999/il-repack/blob/18698fddc4bb577f676d7465a6ad271ced58abc1/ILRepack/ILRepack.cs#L450-L462


~~ Other repositories ~~

We will need to switch to `$(TargetFrameworkVersion)`=v4.7.2 and use
the NUnit adapter in other repos:

  - Downstream in `monodroid`, things will likely break without these
    changes.
  - `Java.Interop` is currently *working*, but should be updated.
  - `xamarin-android-tools` is currently *working*, but should be
    updated.

[0]: https://web.archive.org/web/20190307085042/https://visualstudio.microsoft.com/vs/preview/
[1]: https://github.com/Microsoft/msbuild/blob/e03296fe14152b4c84e6610d2fd2fd10f1ecba82/src/Utilities/ToolLocationHelper.cs#L1736-L1747