[build] Add support for commercial builds
Introduces an '.external' file at the root of the repository that can be
used to track external git dependencies without requiring that they be
added as a submodule.

Entries in this file which start with a '#' will be ignored. Entries
which do not match the specific format will also be ignored. Valid
formats are as follows:

    $(RepoOwner)/$(RepoName):$(BranchName)@$(CommitHash)

This file currently only contains one reference to commercial source,
which is used by the new 'prepare-external-git-dependencies' and
'commercial' make rules. In the future however we may want to extend
this file to reference other dependencies tied to a git repo/branch/hash
(such as a mono bundle).

A 'ParseExternalGitDependencies' task has been added to parse each line
of this file into individual items that can be used by other tasks
and targets.

A 'CheckoutExternalGitDependency' task has been added to clone or update
a local checkout of a source dependency. The Output items from the
'ParseExternalGitDependencies' task should be used as an input to this
task.

The 'external-git-dependencies.targets' file has been added to expose
these common '.external' file processing tasks in MSBuild targets. A
reference to this file has been added to 'xa-prep-tasks.csproj' to allow
us to  easily checkout, download or otherwise process the external git
dependencies specified in the '.external' file. Currently only source
dependencenies named 'monodroid' are supported by these targets.

The 'prepare-external-git-dependencies' make rule is used to ensure our
'.external' dependencies are initialized. The new 'commercial' make rule
has been added to manage the build of commercial components.

Finally, OpenTK.dll and it's related files have been removed from the
installers, as they are no longer going to be built as part of this new
"minimal" commercial build.