[create-pkg] Install `ThirdPartyNotices.txt` on macOS
The `ThirdPartyNotices.txt` file is a comprehensive listing of all 3rd
party code which is included in a Xamarin.Android distribution; see
also commit 2bd13c4a.

Commit 2bd13c4a also added `ThirdPartyNotices.txt` to the `.vsix`
installer, allowing it to be read *without* needing to install the
`.vsix` file on Windows; simply `unzip`ing the `.vsix` would allow it
to be viewed.

Unfortunately, we forgot to add `ThirdPartyNotices.txt` to the macOS
installer, and thus e83ba0dd likewise doesn't install it.

Update `build-tools/create-pkg` so that `ThirdPartyNotices.txt` is
installed on macOS into the path:

	/Library/Frameworks/Xamarin.Android.framework/Versions/Current/ThirdPartyNotices.txt