[sqlite-xamarin] Update ThirdPartyNotice information
The `xamarin/sqlite` repo contains two related-yet-distinct sets of
source code:

 1. The SQLite source code, in which the authors have disclaimed
    copyright, dedicating it to the public domain, and

 2. Google's additions for use on Android, which is licensed under the
    Apache-2.0 license.

We've been omitting (2) in favor of (1), which isn't quite right.

Update `sqlite.tpnitems` so that *both* sources are listed.  This will
cause `sqlite/sqlite` to be listed, along with it's disclaimer of
copyright, alongside `xamarin/sqlite` with the Apache 2.0 license.