[Xamarin.Android.Build.Tasks] fix missing $(_AfterCompileDex)
Continuing the effort to remove `BeforeTargets` and `AfterTargets`, I
am now making changes in monodroid.

One such case is an MSBuild target that needs to run *after*
`_CompileDex`, which is using `AfterTargets` right now.

We already have `$(_AfterCompileDex)`, but it is not specified in
xamarin-android everywhere it needs to be. Anywhere that `_CompileDex`
is listed as a `DependsOnTargets`, we need to list
`$(_AfterCompileDex)` as well.

This change should allow me to finish up the changes in monodroid.