[Xamarin.Android.Build.Tasks] Obsolete OpenTK 0.9.3 (#2852)
We want to remove OpenTK 0.9.3 in the next major release of
Xamarin.Android.

`OpenTK-1.0.dll` should be used instead, and has been shipped with
Xamarin.Android since Mono for Android 4.4.  @jonpryor doesn't
remember when *exactly* that was released, but Xamarin.Android 4.7
was released to alpha in 2013-Mar, so this is *at least* 6 years ago.

Generate an XA1009 warning if anyone is still using `OpenTK.dll`,
so that developers will have a chance to update their projects.