[Xamarin.Android.Build.Tasks] fix missing $(_AfterCompileDex) (#2849)
Continuing the effort to remove `BeforeTargets` and `AfterTargets` --
see commit 39424962 -- I am now making changes in monodroid.

One such case is an MSBuild target that needs to run *after* the
`_CompileDex` target, which is using `AfterTargets` right now.

We already have `$(_AfterCompileDex)`, but it is not used everywhere
in xamarin-android that it needs to be.  Anywhere that `_CompileDex`
is listed as a `DependsOnTargets`, we need to list
`$(_AfterCompileDex)` as well.

This change should allow me to finish up the changes in monodroid.