[Xamarin.Android.Build.Tests] Add support for testing the UI
This commit adds support for testing that the
app launches on device AND can be interacted
with. It does this by making use of a few features
of `adb`.

First is `adb input`. This allow us to send taps and
clicks to the device. It simulates user input. So
we can use this to say click a button.

Next up is the `uiautomator` this allows use to get
the complete ui layout for the active activity.
Using

	exec-out uiautomator dump /dev/tty

we can send this data to the console. As it is
xml we can then parse it looking for our ui element
we want. We can then get the bounds of the control
and pass that information to `adb input tap`.

This commit also adds support for testing that
the timezone set on a device is correctly picked
up by mono.

And finally we add support for testing that the
debugger can actaully connect to the application.
These tests are only run if the commerial parts
of the product are available.

Seeing as all of these new tests REQUIRE the
emulator or device they have been split out into
a new `MSBuildDeviceIntegration` project. This
is built as part of the main solution. However
it is run as part of the `make run-apk-tests`
call. This is so we can be sure the emulator is
running.

In future any MSBuild related tests that REQUIRE
a device should be placed in this new project.