[tests] avoid shared dirs for parallel MSBuild tests
Running our MSBuild NUnit tests can have quite a few random failures
due to shared directories:

* `%UserProfile%\.nuget` contains a system-wide cache of NuGet
  packages. Commonly the MSBuild `Restore` target can fail if two
  process restore the same package at the same time.
* `%LocalAppData%\XamarinBuildDownloadCache` commonly the
  Xamarin.Build.Download NuGet can fail if two processes try to
  dowload files at the same time.

These issues happen on a clean build machine (that has never built
xamarin-android). This can happen on macOS as well, but the paths are
different.

An improvement in our sanity here, would be to configure these
directories so they use a path on a per-test basis.

So for NuGet:

https://docs.microsoft.com/en-us/nuget/consume-packages/managing-the-global-packages-and-cache-folders

We can set the `NUGET_PACKAGES` environment variable to a `.nuget`
directory in the project directory of each test.

For Xamarin.Build.Download:

https://github.com/xamarin/XamarinComponents/blob/5e89e3cc935cea98e7f7ac4cefa6f51ca0262d6f/Util/Xamarin.Build.Download/source/Xamarin.Build.Download/Xamarin.Build.Download.targets#L12

We can set the `$(XamarinBuildDownloadDir)` MSBuild property to an
`xbd` directory.

However, in practice this property required a trailing slash, which
opened a problem with our our response files are generated in tests:

    /p:XamarinBuildDownloadDir="C:\src\xamarin-android\bin\TestDebug\temp\Foo Bar\xbd\"

The `\"` at the end is effectively escaping the `"`, so the property
contains the `"`.

The fix is to escape all `\` characters in this response file:

    /p:XamarinBuildDownloadDir="C:\\src\\xamarin-android\\bin\\TestDebug\\temp\\Foo Bar\\xbd\\"

Now everything works.

I chose short directory names for `.nuget` and `xbd` to avoid
`MAX_PATH` issues.

I also removed `[NonParallelizable]` from several MSBuild tests. We
should only keep them on a few tests that rely on an operation
completing in a length of time.