[Xamarin.Android.Build.Tasks] Add custom metadata to @(ReferencePath) (#2823)
As part of ongoing work with the support libraries, we will want to
start adding custom metadata to the `@(ReferencePath)`.

This data will inform the Xamarin.Android build process that it can
skip processing of certain items.  Or remove that assembly completely
from the final package.

We did some initial work on this in commit 02c07ed6.  However that
was strictly limited to the `<ConvertResourcesCases/>` task.  We now
need a more generic solution.

In this case we have a new `<AppendCustomMetadataToItemGroup/>` task,
which takes two item groups and will copy the metadata from the
`MetaDataItems` items to the `Inputs` items.  This is based on a
`%(Identity)` comparison.  The `Outputs` will be the `Inputs` list
with any additional metadata added to the appropriate items.  We can
then use this list in the build process and query metadata as needed.

The `MetaDataItems` list can contain a number of different
`ItemGroup`s with different metadata.  If this is the case the result
will be a union of the two:

	<ItemGroup>
	  <I Include="Foo" />
	  <A Include="Foo">
	    <Meta1>True</Meta1>
	  </A>
	  <B Include="Foo">
	    <Meta2>False</Meta2>
	  </B>
	</ItemGroup

Given the above the items, the result of:

	<AppendCustomMetadataToItemGroup
	    Inputs="@(I)"
	    MetaDataItems="@(A);@(B)">
	  <Output TaskParameter="Output" ItemName="_NewI" />
	</AppendCustomMetadataToItemGroup>

The `@(_NewI)` item group will be equivalent to:

	<ItemGroup>
	  <I Include="Foo>
	    <Meta1>True</Meta1>
	    <Meta2>False</Meta2>
	  </I>
	</ItemGroup>

Update `Xamarin.Android.Common.targets` to use `@(_ReferencePath)`
instead of `@(ReferencePath)`, wheren `@(_ReferencePath)` copies
`@(ReferencePath)` and adds the metadata from
`@(_AndroidAssemblySkipCases)`.  This allows us to simplify
`<ResolveLibraryProjectImports/>`, as
`%(_ReferencePath.SkipAndroidResourceProcessing)` can be used instead
of separately providing `@(_AndroidAssemblySkipCases)`.