[tests] Rework logcat-*.txt handling (#2857)
Context: https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/231/testReport/Xamarin.Android.Bcl_Tests/xunit/Possible_Crash___Release/

Sometimes an on-device test crashes, but the `adb logcat` output is
*not* included into the NUnit error report that `<RenameTestCases/>`
generates.  This is annoying, as it requires that we download the
`xa-test-results*.zip` file to obtain the actual `adb logcat` output,
which is time-consuming and not entirely obvious.

Additionally, I don't understand *why* the `adb logcat` output isn't
being included; the test execution log states:

	Task "RunInstrumentationTests"
	  ...
	  Task Parameter:NUnit2TestResultsFile=…/xamarin-android/tests/../bin/TestRelease/TestResult-Xamarin.Android.Bcl_Tests.xunit.xml
	  ...
	  Executing: /Users/builder/android-toolchain/sdk/platform-tools/adb -s emulator-5570  shell am instrument  -e "loglevel Verbose" -w "Xamarin.Android.Bcl_Tests/xamarin.android.bcltests.XUnitInstrumentation"
	  INSTRUMENTATION_RESULT: shortMsg=Process crashed.
	  INSTRUMENTATION_CODE: 0
	  Executing: /Users/builder/android-toolchain/sdk/platform-tools/adb -s emulator-5570  logcat -v threadtime -d
	    appending stdout to file: logcat-Release-Xamarin.Android.Bcl_Tests.txt
	  --------- beginning of main

Note the second-to-last line: `appending stdout to file`.

`<RenameTestCases/>` *should* be looking for
`logcat-Release-Xamarin.Android.Bcl_Tests.txt`, the file appended to
above, yet it isn't being found; if it were, it would be in the
`TestResult-Xamarin.Android.Bcl_Tests.xunit.xml` file and part of the
Jenkins-displayed error message!

This "feels like" there's a mismatch in directories being used to
store the `logcat-*` files, which is why they're not found.

To attempt to address this, let's be more specific: instead of relying
on the current-working-directory, provide full path names to the
`logcat-*` files to create, by way of `$(MSBuildThisFileDirectory)`.
Additionally, change the directory that the `logcat-*` files are
stored from `$(topdir)/tests` to `$(topdir)/bin/Test$(Configuration)`.
This prevents "polluting" the `tests` directory and is more
consistent.

Additionally, update `<RenameTestCases/>` so that it logs which
`adb logcat` file it's attempting to read, so that if there *is* a
directory mismatch, we can better diagnose it.

Finally, update `make package-build-status` and
`make package-test-results` so that the `logcat-*` files are actually
packaged.  As per commit e138034e, `make package-build-status` is
executed *before* unit tests are run, meaning there will not be any
`logcat-*` files to include in `make package-build-status`.  Remove
the `logcat-*` files from `make package-build-status`, and update
`make package-test-results` to instead include the `logcat-*` files.