[build] Remove external refs from xa-prep-tasks
We'd like to have an msbuild task project that can be used for preparing
our build environment and providing other general functionality without
taking any external dependencies. In order to accomplish this, the
`JdkInfo` task has been moved into Xamarin.Android.Tools.BootstrapTasks,
which is likely a more fitting home for this task.