[Xamarin.Android.Build.Tools] Bump Android Tools versions. (#2835)
Fixes: https://github.com/xamarin/xamarin-android/issues/2433

We should be using the latest versions of the Android tooling if
possible, matching the major version with the max API level we
support.  In this case we are on Android Pie, API-28.
Thus we should be using the 28.x.x versions of the tooling.

Update `$(AndroidSdkBuildToolsVersion)` to 28.0.3.  
Update `$(AndroidSdkPlatformToolsVersion)` to 28.0.0.

Since not all of the tools have a 28.x.x version, in those cases we
should just use the very latest.