Annotate attributes with resource category hint for the IDE
The code completion experience can be improved if the editor knew
which string properties typically come from resources so a completion
suggestion can be provided that is filtered to the type of resource
the property should be pointing to (i.e. `@(string)` vs `@(drawable)`).

Fixes https://github.com/xamarin/xamarin-android/issues/2862