[Linux build] Debian/testing has libncurses6 instead of libncurses5
We're trying to upgrade our Linux build bots to a newer version of Debian in
hopes to fix a problem with cross-building LLVM for Windows with MinGW. However,
Debian/testing (codename buster) does not install `libncurses5` by default,
instead it comes with `libncurses6` which makes the preparation phase of the
build fail.

This commit adds a check for Debian version and adjusts the package list
accordingly.