[java-runtime] Remove debug files from java_runtime.jar
Commit xxxx added the new java_runtime.jar and
java_runtime_fastdev.jar files. The idea is the
java_runtime.jar was the release version of the
runtime. It should NOT contain files such as

	mono/android/ResourcePatcher.class
	mono/android/Seppuku.class
	mono/android/MonkeyPatcher.class

These are for fast dev only.

However we were including these files!!

This was because of a couple of problems.
1) The `RemoveItems` property of `_RuntimeOutput` was being
treated like an ItemGroup when it is infact a property.
So we need to semi-colon seperate the items to make sure
they are ALL removed rather then it just being the last one
listed.

2) Because the `Inputs` was changed to `@(_RuntimeOutput)`
it means the target will run only ONCE. As a result both the
`release.txt` and `fastdev.txt` files contained the SAME
file list. What we should have been doing was using Target
Batching via

	%(_RuntimeOutput.OutputJar)

This makes sure the target runs for EACH output. This will
result in the behaviour we want, i.e  the files listed
above NOT being in the release version of the jar.