[installers] use @(AndroidApiInfo) for frameworks list (#2868)
Commit e83ba0dd implicitly dealt with a tension: should item groups
*exhaustively* and *explicitly* specify files to include, or should
wildcards be used?

The advantage to explicit lists is that it's easier to manually
determine what files *should* and should *not* be included, and allow
MSBuild to reasonably report an error if a file is *missing*.

The advantage to wildcards or "pattern"-based lists is with
maintenance: it's easier for things to Just Work™ if everything
doesn't meed to be explicitly specified.

There is a tension here, and e83ba0dd attempted to used both, as
deemed appropriate at the time.  As part of that, the framework
assemblies were *mostly* explicitly specified, manually listing
files for frameworks `v4.4` through `v8.1`, then using
`$(AndroidLatestFrameworkVersion)` for the latest framework.

This "broke" when 2cace0b3 changed `$(AndroidLatestFrameworkVersion)`
to `v9.0.99` as part of the API-Q changes, and *didn't* manually
insert the framework files for `v9.0`, resulting in
`Xamarin.Android.Sdk*.vsix` and `.pkg` installers lacking the `v9.0`
framework files:

	# No v9.0 files!
	$ zipinfo ~/Downloads/Xamarin.Android.Sdk-OSS-9.2.99.95_master_58137da.vsix | grep Mono.Android.dll
	-rw----     0.0 fat 17827328 b- defN 19-Mar-22 15:14 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v4.4/Mono.Android.dll
	-rw----     0.0 fat 17888256 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v4.4.87/Mono.Android.dll
	-rw----     0.0 fat 19628544 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v5.0/Mono.Android.dll
	-rw----     0.0 fat 19720704 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v5.1/Mono.Android.dll
	-rw----     0.0 fat 20758528 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v6.0/Mono.Android.dll
	-rw----     0.0 fat 22979584 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v7.0/Mono.Android.dll
	-rw----     0.0 fat 23044096 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v7.1/Mono.Android.dll
	-rw----     0.0 fat 24596480 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v8.0/Mono.Android.dll
	-rw----     0.0 fat 24669184 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v8.1/Mono.Android.dll
	-rw----     0.0 fat 26605568 b- defN 19-Mar-22 15:15 $ReferenceAssemblies/Microsoft/Framework/MonoAndroid/v9.0.99/Mono.Android.dll

In retrospect, requiring that an `$(AndroidLatestFrameworkVersion)`
change also alter the installers may not have been a good choice
on the "explicit vs. pattern-or-wildcard" spectrum.

Update `build-tools/installers/create-installers.targets` so that
`@(AndroidApiInfo)` and a directory existence check is used to
construct the list of files to include.  This ensures that whenever
`src/Mono.Android/Mono.Android.projitems` is updated -- which
declares `@(AndroidApiInfo)` -- the installer will automatically pick
up the appropriate files.

Unfortunately, getting the MSBuild transform right was quite tricky...

At first, I just tried updating the existing `<ItempGroup/>`:

	<ItemGroup>
	  <_FrameworkDirs Include="@(AndroidApiInfo->'$(FrameworkSrcDir)\%(Identity)')" />
	  <_FrameworkDirsThatExist Condition="Exists('%(Identity)')" Include="@(_FrameworkDirs)" />
	  <_FrameworkFiles Include="@(_FrameworkDirsThatExist->'%(Identity)\AndroidApiInfo.xml')" />
	  <_FrameworkFiles Include="@(_FrameworkDirsThatExist->'%(Identity)\mono.android.dex')" />
	  <_FrameworkFiles Include="@(_FrameworkDirsThatExist->'%(Identity)\Mono.Android.dll')" />
	  <_FrameworkFiles Include="@(_FrameworkDirsThatExist->'%(Identity)\mono.android.jar')" />
	  <_FrameworkFiles Include="@(_FrameworkDirsThatExist->'%(Identity)\Mono.Android.pdb')" />
	  <_FrameworkFiles Include="@(_FrameworkDirsThatExist->'%(Identity)\RedistList\FrameworkList.xml')" />
	</ItemGroup>

This construct fails with:

    error MSB4190: The reference to the built-in metadata "Identity" at position 8 is not allowed in this condition "Exists('%(Identity)')".

However, it works when moved *inside* a target?  I am guessing that
`%(Identity)` does not exist at evaluation time.

MSBuild is a weird thing.

I just put this `<ItemGroup/>` inside the `ConstructInstallerItems`
MSBuild target.

This should make things future proof for new
`$(TargetFrameworkVersion)` values, while also allowing PR builds
to work.