[Xamarin.Android.Build.Tasks] fix for <Aapt/>
Context: http://build.devdiv.io/2527883

Our integration build for the Xamarin designer has caught a bug in the
following situation:

* macOS-only (Windows is OK)
* `$(AndroidUseAapt2)` is `False`
* Using `Xamarin.Android.Support.CustomTabs` version 23.x

I was able to reproduce the problem by changing the
`BuildTest.ResourceExtraction` test.

The error message is:

    /Library/Frameworks/Mono.framework/External/xbuild/Xamarin/Android/Xamarin.Android.Common.targets(1765,2):
        error APT0000: resource directory '/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/ADT-0/RT-MyDriving-1/MyDriving.Android/obj/Debug/resourcecache/3288F845948C6081AB114E82B328E204/res' does not exist

The reason this fails is that this particular library does not have a
`res` directory:

    $ ls -l ~/.local/share/Xamarin/Xamarin.Android.Support.CustomTabs/23.1.1.0/embedded
    total 64
     8 -rw-r--r--  1 jonathanpeppers  staff    861 Mar 26 11:18 AndroidManifest.xml
     0 drwxr-xr-x  3 jonathanpeppers  staff     96 Mar 26 11:18 aapt
    56 -rw-r--r--  1 jonathanpeppers  staff  24743 Mar 26 11:18 classes.jar

The fix is to add a `Directory.Exists` check, this seems to fix the
unit test *and* the failing project used by the designer build.