[Xamarin.Android.Build.Tasks] Check a jar has .class files in it.
We have found an issue with our fastdev support when
it comes across a `.jar` file which does NOT contain
any `.class` files.

As part of our fastdev system we convert all the
`.jar` files we need to `.dex` files so they
can be fast deployed.

If we come across a `.jar` file which does not
contain any `.class` files `dx` exists with the
following error

	no classes!

This causes the build to fail. So what we need to do
is check that a `.jar` has `.class` files in it.
This commit also reworks some code to not use Linq
in favour of a standard foreach loop.