[java-runtime] Remove `final` from MonoPackageManager_Resources (#2878)
In commit 75023bd2 we noticed that the value of
`MonoPackageManager_Resouces.ApiPackageName` was `""` even after it
was replaced at build time with a new `.class` file which contained
an actual value:

> We don't yet fully understand this failure scenario, but setting
> `ApiPackageName` to null allows things to work.

The fix in 75023bd2 was to make the default value within
`MonoPackageManager_Resources.java` have a value of `null` rather
than `""`.  This allowed things to work but we didn't know WHY.

After some more investigation it turns out that when
`MonoPackageManager_Resouces.ApiPackageName` was `""`, the empty
string value was being *inlined* by `javac` when generating
`MonoPackageManager.getApiPackageName()`:

	public static java.lang.String getApiPackageName();
	    descriptor: ()Ljava/lang/String;
	    Code:
	       0: ldc           #48                 // String
	       2: areturn

`ldc` loads a constant from the Constants table, in this case `""`.

When `null` was used, `javac` instead emitted a field access:

	public static java.lang.String getApiPackageName();
	    descriptor: ()Ljava/lang/String;
	    Code:
	       0: getstatic     #47                 // Field mono/MonoPackageManager_Resources.ApiPackageName:Ljava/lang/String;
	       3: areturn

(Why `javac` didn't inline `null` is beyond @jonpryor.)

This is why changing to `null` worked: it prevented the compiler from
inlining the value.

However, while that "worked," it is not at all obvious, and short of
reading the Java Language Spec to determine if `null` is even a value
which *can* be inlined, the current approach of using a `final` value
seems potentially *dangerous*, as behavior may be subject to the
compiler used and subject to change.

To fix this, remove the `final` keyword from the
`MonoPackageManager_Resources` fields prevent inlining.