[Xamarin.Android.Build.Tests] Add code to get the device abi
Currently all of our unit tests assume that we
are testing against an x86 device. This commit
adds code to get the abi of the attached device
to that it is available for the unit tests.

Currently there are no tests in this repo which
will use this. However we do have a plan to move
all our tests here and there are a few fastdev
tests which weill make use of this.