[mono-runtimes] Mono Archive Support (#2817)
Build `src/mono-runtimes` using the [Mono Archive][0] (f970cd50), and
update `@(MonoProfileAssembly)` and `@(MonoTestAssembly)`.

A "funny" thing happened around mono archive support: it didn't work.
Specifically, mono archive support was never added to the
[mono/2018-08 branch][1].  It *was* added to the
[mono/2018-10 branch][2], but mono/mono@dc977160 altered the filename
format, preventing `src/mono-runtimes` from using it.

There was an [attempt to fix xamarin-android][3], but in the interest
of "sanity" and "not doing too much at once," it was [removed][4]
from the [mono/2018-10 bump PR][5]:

> Is it necessary to intermix the "mono archives" changes along with
> the mono/2018-10 bump changes?
> 
> I mean sure, mono/2018-10 has improved support for the mono
> archive, but I fear/suspect that the mono archive changes
> [are causing the NUnit tests to crash][6], and thus a smaller, more
> focused PR may be more desirable (and then deal with the mono
> archive changes in a separate PR).

Now that [xamarin-android is using mono/2018-10][7], it is time to
revisit mono archive support:

 1. Fix the mono archive URL that we attempt to download, so that
    we can actually *find* a mono archive.

 2. Fix `src/mono-runtimes` so that we can *build* with the archive.

 3. Fix BCL unit test resource usage.

 4. Update `@(MonoProfileAssembly)` and `@(MonoTestAssembly)`.

 5. Improve unit test name exclusion behavior


~~ Build fixes ~~

Many of the paths that `src/mono-runtimes` looked for were wrong when
using the mono archive, e.g. using `external/mono/sdks/builds` when
that directory is used for building mono, but when the archive is
used that directory is empty.  Update those paths to instead use
`external/mono/sdks/out`, which is *also* valid for mono source
builds *and* for mono archive use.

Similarly, `Consts.cs` is a generated file which is included in the
mono archive, but we needed to update our path to use it from
`external/mono/sdks/out` and not the "normal" build location.

Finally, use the `<SystemUnzip/>` task to extract the mono archive
`.zip` file.  This at least has a *plausible* chance of working on
Windows, should we wish to pursue Windows-native support of mono
archives.


~~ Unit test resources ~~

The `tests/BCL-Tests`-related projects execute mono-built unit test
assemblies, which are included in the mono archive.  When we first
added support for executing the mono-provided unit test assemblies
in e9daf5ea, we also included "unit test resources" for extraction
on the Android device.  These unit test resources were files that the
unit test expected to exist as files on the filesystem.

The mono unit tests have since been fixed so that these extra
resources are now included within the unit test assemblies as
`@(EmbeddedResource)`s.  There is thus no longer a need to separately
package and extract  `@(_BclTestContent)`.  This allows us to remove
and simplify `mono-runtimes.targets`.  (Yay!)

One ramification of this change is around *satellite assemblies*:
`Xamarin.Android.Bcl-Tests.targets` previously would execute the
`<AL/>` task in order to compile certain resources into satellite
assemblies.  This is no longer necessary, as the satellite assemblies
are included in the mono archive.  However, we do need a way of
specifying what the satellite assemblies are; introduce a new
`@(MonoTestSatelliteAssembly)` item group to specify them.


~~ Update `@(MonoProfileAssembly)` and `@(MonoTestAssembly)` ~~

Once we had things building, a possible mistake was made: we updated
the `@(MonoProfileAssembly)` and `@(MonoTestAssembly)` item groups
to include what was in the mono archive.

Add a new `build-tools/scripts/gen-ProfileAssemblies.sh` script which
will generate the contents of
`src/mono-runtimes/ProfileAssemblies.projitems` based on the contents
of `external/mono/sdks/out`, which is where the mono archive is
extracted.  This will generate appropriate `@(MonoFacadeAssembly)`,
`@(MonoProfileAssembly)`, `@(MonoTestAssembly)`, and
`@(MonoTestSatelliteAssembly)` item groups:

	build-tools/scripts/gen-ProfileAssemblies.sh > src/mono-runtimes/ProfileAssemblies.projitems

On the plus side, updating these item groups needed to be done
*anyway*: if there are new unit tests, we should execute them!
(This should have been done as part of the mono bumps, but it's
easily overlooked.)


~~ Unit test exclusion ~~

Unfortunately, not all of the new unit tests pass: `System.IO.Pipes`
doesn't support Android or Linux:

	[$xUnit] Test FAILED [monodroid_System.Core_xunit-test, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null]
	   Test name: System.IO.Pipes.Tests.AnonymousPipeTest_CreateClient.CreateClientStreamFromStringHandle_Valid
	   Exception messages: System.PlatformNotSupportedException : Operation is not supported on this platform.
	   Exception stack traces:   at System.IO.Pipes.AnonymousPipeServerStream.Create (System.IO.Pipes.PipeDirection direction, System.IO.HandleInheritability inheritability, System.Int32 bufferSize)

The `MonoTests.System.Web.Services.Description.ServiceDescriptionTest`
unit tests also fail:

	System.TypeInitializationException : The type initializer for 'System.Web.Services.Description.ExtensionManager' threw an exception.
	  ----> System.InvalidOperationException : There was an error reflecting type 'System.Web.Services.Description.HttpAddressBinding'.
	  ----> System.InvalidOperationException : System.Web.Services.Description.HttpAddressBinding cannot be serialized because it does not have a default public constructor

Additionally, several of the added `System.Linq.Expressions` tests
are not "linker safe" and would fail in Release builds but not Debug
builds.

Finally, we discovered that our unit test name exclusion strategy in
`xunit-excluded-tests.txt` was insufficient, as when xUnit has
parameterized tests, those parameters become part of the test name:

	System.IO.Pipes.Tests.AnonymousPipeTest_Read_ServerIn_ClientOut.AsyncReadWriteChain_ReadWrite(iterations: 5000, writeBufferSize: 1, readBufferSize: 1, cancelableToken: True)

I don't want to have to repeat every one of those parameterized
variations, as it could quickly become out of date, and determining
a complete list would be harder.

Improve `XUnitTestRunner.IsIncluded()` twofold:

 1. Only require `testCase.Traits` when performing
    `XUnitFilterType.Trait` filtering.  (Some unit tests *lack*
    traits, and without this change we'd never get around to test
    name checks.)

 2. Improve `XUnitFilterType.TypeName` checking so that if a test
    name starst with an excluded name followed by a `(`, the test is
    also skipped.  This allows excluding parameterized tests by
    providing only the name before the `(`.

Exclude all of the tests which currently fail in Release builds.

[0]: https://github.com/xamarin/xamarin-android/projects/10
[1]: https://github.com/mono/mono/commits/2018-08
[2]: https://github.com/mono/mono/commits/2018-10
[3]: https://github.com/xamarin/xamarin-android/pull/2366
[4]: https://github.com/xamarin/xamarin-android/pull/2253#discussion_r246627245
[5]: https://github.com/xamarin/xamarin-android/pull/2253
[6]: https://github.com/xamarin/xamarin-android/pull/2253#issuecomment-452954145
[7]: https://github.com/xamarin/xamarin-android/commit/d87d00f4f8e6f5753dc90a4762d3cdae2a535b5b