[Xamarin.Android.Build.Tests] Add framework for Commercial Builds
We want to move all of our unit tests into one place.
This means that tests now need to know if they are
running against a Open Source build or a Commercial
build.

We detect that by looking for the
`Xamarin.Android.Common.Debugging.targets` in the
Android MSBuild directory.

We also change over the TestCaseSource in the
`BuildTest.OSS.cs` to be methods rather than
properties. This is so we can (in the future)
dynamically update the items to test both the
Open Source and Commercial build behavour.
There are differences between the two builds
(like fastdev) so the test cases are not always
the same.

We also updated ALL the `TestCaseSource` attributes
to name use of `nameof` rather than strings.